Modified to allow for n threads
I have now modified this program to use a parallel map allowing for n threads. The range to be check is provided in the form of a list (e.g. [1..10000]). That list is then broken into chunks using the splitEvery function. (splitEvery :: n -> [a] -> [[a]]). Next, parMap is used to evaluate the list of lists in parallel. Each list is run through a function which filters by isPrime, returning a sublist containing only prime numbers. At this point the result is a list of lists of prime numbers ([[Int]]).

Issues:
- The returned value is a list of lists instead of a list of ints. This can be fixed by concating all the lists together.
- The partition size is fixed at 500000, regardless of the number of cores.
      > I can't find an easy way to determine the number of cores from
      > inside the program. In the future I may modify the program so
      > the partition size always results in n threads, where n is
      > either the number of cores or an external variable.