Now accepts (and expects) two command line args.
The proper way to run this is now "./main 1 1000 +RTS -N" where 1 is the
lower limit and 1000 is the upper limit. -N ensures it will
automatically use as many threads as you computer has cores.
Hyperthreaded "cores" count in this case and your performance may
decrease if more threads are spawned than there are physical cores.